Avoid using hostssl without certs
The default value is a dangerous one as a drop-in replacement for the official postgres image. It produces these errors:

```
LOG:  hostssl requires SSL to be turned on
HINT:  Set ssl = on in postgresql.conf.
CONTEXT:  line 10 of configuration file "/etc/postgres/pg_hba.conf"
FATAL:  could not load pg_hba.conf
LOG:  database system is shut down
```

To avoid that problem, no `hostssl` entries are created if there are no certs.